Refactor; extract graph build step.
